Was converting points to integers. Now fixed.

+++ FileReader.h
@@ -88,7 +88,7 @@ void FileReader::ReadFile()
 
         // Store the data
         std::vector<double> points;
-        points.push_back(std::stoi(words.at(0))); points.push_back(std::stoi(words.at(1))); points.push_back(std::stoi(words.at(2)));
+        points.push_back(std::stod(words.at(0))); points.push_back(std::stod(words.at(1))); points.push_back(std::stod(words.at(2)));
         if (isClustering) this_cluster.push_back(points);
         tpc_event.data.push_back(points);
       } // while getline
